Tp-MT-1/requirements.txt
autopep8==2.0.1
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.6
filelock==3.8.0
fonttools==4.51.0
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.8.4
numba==0.59.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
platformdirs==2.5.2
pycodestyle==2.10.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.1
virtualenv==20.16.6
